Error checking in populateBuffer!
Added an option to raise exceptions if received packet metadata
indicates an error (see #10). This is disabled by default if the
streamer is configured in continuous mode, to avoid breaking existing
code that can ignore overflow errors and doesn't care about the exact
number of samples received.

The exceptions are not exported to avoid adding generic names such as
OverflowException to the global namespace.